+++ app/gui/src/project-view/components/GraphEditor/GraphNode/nodeVisualization.ts
@@ -1,7 +1,7 @@
 import type GraphVisualization from '@/components/GraphEditor/GraphVisualization.vue'
 import { type RawDataSource } from '@/components/GraphEditor/GraphVisualization/visualizationData'
-import { injectKeyboard } from '@/providers/keyboard'
-import { TypeInfo } from '@/stores/project/computedValueRegistry'
+import { injectBubblingKeyboard } from '@/providers/keyboard'
+import type { TypeInfo } from '@/stores/project/computedValueRegistry'
 import { type VisualizationDataSource } from '@/stores/visualization'
 import { type Opt } from '@/util/data/opt'
 import { type Rect } from '@/util/data/rect'
@@ -43,7 +43,7 @@ export function useNodeVisualization({
   hidden,
   emit,
 }: NodeVisualizationOptions) {
-  const keyboard = injectKeyboard()
+  const keyboard = injectBubblingKeyboard()
   const metadata = computed(() => toValue(vis))
   const visualizationWidth = computed<number | null>({
     get: () => metadata.value?.width ?? null,