+++ engine/runtime/src/main/scala/org/enso/interpreter/runtime/IrToTruffle.scala
@@ -14,6 +14,7 @@ import org.enso.compiler.core.ConstantsNames
 import org.enso.compiler.core.Implicits.AsMetadata
 import org.enso.compiler.core.IR
 import org.enso.compiler.core.ir.{
+  AscriptionReason,
   CallArgument,
   DefinitionArgument,
   Empty,
@@ -24,7 +25,6 @@ import org.enso.compiler.core.ir.{
   Module,
   Name,
   Pattern,
-  `type`,
   Type => Tpe
 }
 import org.enso.compiler.core.ir.module.scope.Definition
@@ -57,7 +57,6 @@ import org.enso.compiler.pass.resolve.{
   GenericAnnotations,
   GlobalNames,
   Patterns,
-  TypeNames,
   TypeSignatures
 }
 import org.enso.interpreter.node.callable.argument.ReadArgumentNode
@@ -98,8 +97,10 @@ import org.enso.interpreter.runtime.callable.{
   Annotation => RuntimeAnnotation
 }
 import org.enso.interpreter.runtime.data.Type
-import org.enso.interpreter.runtime.scope.{ImportExportScope, ModuleScope}
+import org.enso.interpreter.runtime.scope.ImportExportScope
+import org.enso.interpreter.runtime.scope.ModuleScopeBuilder
 import org.enso.interpreter.{Constants, EnsoLanguage}
+import org.enso.interpreter.runtime.builtin.Builtins
 
 import java.math.BigInteger
 import java.util.function.Supplier
@@ -126,9 +127,12 @@ import scala.jdk.OptionConverters._
 class IrToTruffle(
   val context: EnsoContext,
   val source: Source,
-  val scopeBuilder: ModuleScope.Builder,
+  val scopeBuilder: ModuleScopeBuilder,
   val compilerConfig: CompilerConfig
 ) {
+  private def getBuiltins: Builtins = {
+    Builtins.get(context)
+  }
 
   val language: EnsoLanguage = context.getLanguage
 
@@ -257,42 +261,46 @@ class IrToTruffle(
           frameInfo
         )
 
-        val function = conversion.body match {
-          case fn: Function =>
-            val bodyBuilder =
-              new expressionProcessor.BuildFunctionBody(
-                conversion.methodName.name,
-                fn.arguments,
-                fn.body,
-                TypeCheckValueNode.single("conversion", toType),
-                None,
-                true
+        val function: Supplier[RuntimeFunction] = () =>
+          conversion.body match {
+            case fn: Function =>
+              val bodyBuilder =
+                new expressionProcessor.BuildFunctionBody(
+                  conversion.methodName.name,
+                  fn.arguments,
+                  fn.body,
+                  TypeCheckValueNode.single(
+                    AscriptionReason.forFunctionResult("conversion"),
+                    toType
+                  ),
+                  None,
+                  true
+                )
+              val rootNode = MethodRootNode.build(
+                language,
+                expressionProcessor.scope,
+                scopeBuilder.asModuleScope(),
+                () => bodyBuilder.bodyNode(),
+                makeSection(scopeBuilder.getModule, conversion.location),
+                toType,
+                conversion.methodName.name
               )
-            val rootNode = MethodRootNode.build(
-              language,
-              expressionProcessor.scope,
-              scopeBuilder.asModuleScope(),
-              () => bodyBuilder.bodyNode(),
-              makeSection(scopeBuilder.getModule, conversion.location),
-              toType,
-              conversion.methodName.name
-            )
-            val callTarget = rootNode.getCallTarget
-            val arguments  = bodyBuilder.args()
-            val funcSchema = FunctionSchema
-              .newBuilder()
-              .argumentDefinitions(arguments: _*)
-              .build()
-            new RuntimeFunction(
-              callTarget,
-              null,
-              funcSchema
-            )
-          case _ =>
-            throw new CompilerError(
-              s"Conversion bodies must be functions at the point of codegen (conversion $fromType to $toType)."
-            )
-        }
+              val callTarget = rootNode.getCallTarget
+              val arguments  = bodyBuilder.args()
+              val funcSchema = FunctionSchema
+                .newBuilder()
+                .argumentDefinitions(arguments: _*)
+                .build()
+              new RuntimeFunction(
+                callTarget,
+                null,
+                funcSchema
+              )
+            case _ =>
+              throw new CompilerError(
+                s"Conversion bodies must be functions at the point of codegen (conversion $fromType to $toType)."
+              )
+          }
         scopeBuilder.registerConversionMethod(toType, fromType, function)
       }
     }
@@ -359,7 +367,7 @@ class IrToTruffle(
     override protected def processTypeDefinition(typ: Definition.Type): Unit = {
       val atomDefs = typ.members
       val asType =
-        scopeBuilder.asModuleScope().getType(typ.name.name, true)
+        scopeBuilder.getType(typ.name.name, true)
       val atomConstructors =
         atomDefs.map(cons => asType.getConstructors.get(cons.name.name))
       atomConstructors
@@ -723,7 +731,7 @@ class IrToTruffle(
 
     val staticWrapper = methodDef.isStaticWrapperForInstanceMethod
 
-    val builtinFunction = context.getBuiltins
+    val builtinFunction = getBuiltins
       .getBuiltinFunction(
         methodOwnerName,
         methodName,
@@ -740,7 +748,7 @@ class IrToTruffle(
       .left
       .flatMap { l =>
         // Builtin Types Number and Integer have methods only for documentation purposes
-        val number = context.getBuiltins.number()
+        val number = getBuiltins.number()
         val ok =
           staticWrapper && (cons == number.getNumber.getEigentype || cons == number.getInteger.getEigentype) ||
           !staticWrapper && (cons == number.getNumber             || cons == number.getInteger)
@@ -803,73 +811,14 @@ class IrToTruffle(
   // === Utility Functions ====================================================
   // ==========================================================================
 
-  private def extractAscribedType(
-    comment: String,
-    t: Expression
-  ): TypeCheckValueNode = t match {
-    case u: `type`.Set.Union =>
-      val oneOf = u.operands.map(extractAscribedType(comment, _))
-      if (oneOf.contains(null)) {
-        null
-      } else {
-        val arr: Array[TypeCheckValueNode] = oneOf.toArray
-        TypeCheckValueNode.oneOf(comment, arr: _*)
-      }
-    case i: `type`.Set.Intersection =>
-      TypeCheckValueNode.allOf(
-        comment,
-        extractAscribedType(comment, i.left),
-        extractAscribedType(comment, i.right)
-      )
-    case p: Application.Prefix => extractAscribedType(comment, p.function)
-    case _: Tpe.Function =>
-      TypeCheckValueNode.single(
-        comment,
-        context.getTopScope().getBuiltins().function()
-      )
-    case typeWithError: Tpe.Error =>
-      // When checking a `a ! b` type, we ignore the error part as it is only used for documentation purposes and is not checked.
-      extractAscribedType(comment, typeWithError.typed)
-    case typeInContext: Tpe.Context =>
-      // Type contexts aren't currently really used. But we should still check the base type.
-      extractAscribedType(comment, typeInContext.typed)
-    case err: errors.Resolution =>
-      TypeCheckValueNode.fail("unresolved symbol " + err.originalName.name)
-    case t => {
-      val res = t.getMetadata(TypeNames)
-      res match {
-        case Some(
-              BindingsMap
-                .Resolution(binding @ BindingsMap.ResolvedType(_, _))
-            ) =>
-          val typeOrAny = asType(binding)
-          if (context.getBuiltins().any() == typeOrAny) {
-            null
-          } else {
-            TypeCheckValueNode.single(comment, typeOrAny)
-          }
-        case Some(
-              BindingsMap
-                .Resolution(BindingsMap.ResolvedPolyglotSymbol(mod, symbol))
-            ) =>
-          TypeCheckValueNode.meta(
-            comment,
-            asScope(
-              mod.unsafeAsModule().asInstanceOf[TruffleCompilerContext.Module]
-            ).getPolyglotSymbolSupplier(symbol.name)
-          )
-        case _ => null
-      }
-    }
-  }
-
   private def checkAsTypes(
     arg: DefinitionArgument
   ): TypeCheckValueNode = {
-    val comment = "`" + arg.name.name + "`"
     arg.ascribedType
       .map { t =>
-        TypeCheckValueNode.allTypes(false, extractAscribedType(comment, t))
+        val reason    = AscriptionReason.forParameter(arg.name.name)
+        val checkNode = IrTruffleUtils.extractAscribedType(context, reason, t)
+        TypeCheckValueNode.allTypes(false, checkNode)
       }
       .getOrElse(null)
   }
@@ -1117,10 +1066,12 @@ class IrToTruffle(
                         s"Source type should be defined in module ${module.getName}"
                       )
                       val conversionFun =
-                        actualScope.lookupConversionDefinition(
-                          sourceTp,
-                          targetTp
-                        )
+                        actualScope
+                          .asModuleScope()
+                          .lookupConversionDefinition(
+                            sourceTp,
+                            targetTp
+                          )
                       org.enso.common.Asserts.assertInJvm(
                         conversionFun != null,
                         s"Conversion method `$conversionMethod` should be defined in module ${module.getName}"
@@ -1238,7 +1189,11 @@ class IrToTruffle(
           processCase(caseExpr, subjectToInstrumentation)
         case asc: Tpe.Ascription =>
           val checkNode =
-            extractAscribedType(asc.comment.orNull, asc.signature)
+            IrTruffleUtils.extractAscribedType(
+              context,
+              asc.reason,
+              asc.signature
+            )
           if (checkNode != null) {
             val body = run(asc.typed, binding, subjectToInstrumentation)
             TypeCheckValueNode.wrap(body, checkNode)
@@ -1267,7 +1222,11 @@ class IrToTruffle(
             ir.getMetadata(TypeSignatures)
           types.foreach { tpe =>
             val checkNode =
-              extractAscribedType(tpe.comment.orNull, tpe.signature)
+              IrTruffleUtils.extractAscribedType(
+                context,
+                tpe.reason,
+                tpe.signature
+              )
             if (checkNode != null) {
               runtimeExpression =
                 TypeCheckValueNode.wrap(runtimeExpression, checkNode)
@@ -1345,7 +1304,7 @@ class IrToTruffle(
     def processType(value: Tpe): RuntimeExpression = {
       setLocation(
         ErrorNode.build(
-          context.getBuiltins
+          getBuiltins
             .error()
             .makeSyntaxError(
               "Type operators are not currently supported at runtime"
@@ -1393,7 +1352,7 @@ class IrToTruffle(
 
             val message = invalidBranches.map(_.message).mkString(", ")
 
-            val error = context.getBuiltins
+            val error = getBuiltins
               .error()
               .makeCompileError(message)
 
@@ -1482,13 +1441,13 @@ class IrToTruffle(
                     ) =>
                   val atomCons =
                     asType(tp).getConstructors.get(cons.name)
-                  val r = if (atomCons == context.getBuiltins.bool().getTrue) {
+                  val r = if (atomCons == getBuiltins.bool().getTrue) {
                     BooleanBranchNode.build(
                       true,
                       branchCodeNode.getCallTarget,
                       branch.terminalBranch
                     )
-                  } else if (atomCons == context.getBuiltins.bool().getFalse) {
+                  } else if (atomCons == getBuiltins.bool().getFalse) {
                     BooleanBranchNode.build(
                       false,
                       branchCodeNode.getCallTarget,
@@ -1509,7 +1468,7 @@ class IrToTruffle(
                     ) =>
                   val tpe =
                     asType(binding)
-                  val polyglot = context.getBuiltins.polyglot
+                  val polyglot = getBuiltins.polyglot
                   val branchNode = if (tpe == polyglot) {
                     PolyglotBranchNode.build(
                       tpe,
@@ -1967,7 +1926,7 @@ class IrToTruffle(
             ConstantObjectNode.build(t)
           } else {
             ErrorNode.build(
-              context.getBuiltins
+              getBuiltins
                 .error()
                 .makeSyntaxError(
                   s"Type for $tp is null"
@@ -2054,47 +2013,47 @@ class IrToTruffle(
         case Error.InvalidIR(_, _) =>
           throw new CompilerError("Unexpected Invalid IR during codegen.")
         case err: errors.Syntax =>
-          context.getBuiltins
+          getBuiltins
             .error()
             .makeSyntaxError(err.message(fileLocationFromSection))
         case err: errors.Redefined.Binding =>
-          context.getBuiltins
+          getBuiltins
             .error()
             .makeCompileError(err.message(fileLocationFromSection))
         case err: errors.Redefined.Method =>
-          context.getBuiltins
+          getBuiltins
             .error()
             .makeCompileError(err.message(fileLocationFromSection))
         case err: errors.Redefined.MethodClashWithAtom =>
-          context.getBuiltins
+          getBuiltins
             .error()
             .makeCompileError(err.message(fileLocationFromSection))
         case err: errors.Redefined.Conversion =>
-          context.getBuiltins
+          getBuiltins
             .error()
             .makeCompileError(err.message(fileLocationFromSection))
         case err: errors.Redefined.Type =>
-          context.getBuiltins
+          getBuiltins
             .error()
             .makeCompileError(err.message(fileLocationFromSection))
         case err: errors.Redefined.SelfArg =>
-          context.getBuiltins
+          getBuiltins
             .error()
             .makeCompileError(err.message(fileLocationFromSection))
         case err: errors.Redefined.Arg =>
-          context.getBuiltins
+          getBuiltins
             .error()
             .makeCompileError(err.message(fileLocationFromSection))
         case err: errors.Unexpected.TypeSignature =>
-          context.getBuiltins
+          getBuiltins
             .error()
             .makeCompileError(err.message(fileLocationFromSection))
         case err: errors.Resolution =>
-          context.getBuiltins
+          getBuiltins
             .error()
             .makeCompileError(err.message(fileLocationFromSection))
         case err: errors.Conversion =>
-          context.getBuiltins
+          getBuiltins
             .error()
             .makeCompileError(err.message(fileLocationFromSection))
         case _: errors.Pattern =>
@@ -2114,7 +2073,7 @@ class IrToTruffle(
       * @return the Nothing builtin
       */
     private def processEmpty(): RuntimeExpression = {
-      LiteralNode.build(context.getBuiltins.nothing())
+      LiteralNode.build(getBuiltins.nothing())
     }
 
     /** Processes function arguments, generates arguments reads and creates
@@ -2335,7 +2294,7 @@ class IrToTruffle(
         case _: Application.Typeset =>
           setLocation(
             ErrorNode.build(
-              context.getBuiltins
+              getBuiltins
                 .error()
                 .makeSyntaxError(
                   "Typeset literals are not yet supported at runtime"
@@ -2588,9 +2547,9 @@ class IrToTruffle(
       }
   }
 
-  private def asScope(module: CompilerContext.Module): ModuleScope = {
+  private def asScope(module: CompilerContext.Module): ModuleScopeBuilder = {
     val m = org.enso.interpreter.runtime.Module.fromCompilerModule(module)
-    m.getScope()
+    m.getScopeBuilder()
   }
 
   private def asType(
@@ -2598,14 +2557,16 @@ class IrToTruffle(
   ): Type = {
     val m = org.enso.interpreter.runtime.Module
       .fromCompilerModule(typ.module.unsafeAsModule())
-    m.getScope().getType(typ.tp.name, true)
+    val sb = m.getScopeBuilder()
+    sb.getType(typ.tp.name, true)
   }
 
   private def asAssociatedType(
     module: CompilerContext.Module
   ): Type = {
-    val m = org.enso.interpreter.runtime.Module.fromCompilerModule(module)
-    m.getScope().getAssociatedType()
+    val m  = org.enso.interpreter.runtime.Module.fromCompilerModule(module)
+    val sb = m.getScopeBuilder()
+    sb.getAssociatedType()
   }
 
   private def scopeAssociatedType =