+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Next Next Release
+
+#### Enso Language & Runtime
+
+- [Open type check `Type&Any` lets all visible types thru][13225]
+
+[13225]: https://github.com/enso-org/enso/pull/13225
+
 # Next Release
 
 #### Enso IDE
@@ -18,6 +26,8 @@
 - [Add keyboard shortcuts for formatting documentation][13134]
 - [New right-side panel][13135], unified between tabs.
 - [Allow selecting expected types for arguments of grouped components.][13161]
+- [Component documentation now uses Markdown instead of a custom Markdown-like
+  format][13203]
 - [Methods for ‘intersection’ types are now visible in Component
   Browser.][13266]
 - [Allow marking grouped component arguments as required or providing a default
@@ -37,6 +47,8 @@
 - [Project and Setting tab may be now closed with shortcut][13498][13604]. On
   Windows/Linux <kbd>Ctrl</kbd>+<kbd>W</kbd> or <kbd>Ctrl</kbd> + <kbd>F4</kbd>;
   on macOS: <kbd>⌘</kbd> + <kbd>W</kbd>.
+- [Improved Graph Editor's Top Bar responsiveness to width change.][13726]
+- [Changed animation of tabs highlight.][13726]
 
 [12774]: https://github.com/enso-org/enso/pull/12774
 [12778]: https://github.com/enso-org/enso/pull/12778
@@ -52,6 +64,7 @@
 [13134]: https://github.com/enso-org/enso/pull/13134
 [13135]: https://github.com/enso-org/enso/pull/13135
 [13161]: https://github.com/enso-org/enso/pull/13161
+[13203]: https://github.com/enso-org/enso/pull/13203
 [13266]: https://github.com/enso-org/enso/pull/13266
 [13254]: https://github.com/enso-org/enso/pull/13254
 [13347]: https://github.com/enso-org/enso/pull/13347
@@ -66,6 +79,7 @@
 [13479]: https://github.com/enso-org/enso/pull/13479
 [13498]: https://github.com/enso-org/enso/pull/13498
 [13604]: https://github.com/enso-org/enso/pull/13604
+[13726]: https://github.com/enso-org/enso/pull/13726
 
 #### Enso Standard Library